+++ client/go/internal/vespa/document/http.go
@@ -1,4 +1,4 @@
-package feed
+package document
 
 import (
 	"bytes"
