+++ client/go/internal/vespa/document/http_test.go
@@ -1,4 +1,4 @@
-package feed
+package document
 
 import (
 	"bytes"
